Se creó el archivo ecomerce-maria haro que contiene los get, set y string de productos, pedidos  y usuarios
