--- ecomercemariaharo/src/main/java/com/uppoteam/ecomercemariaharo/controller/ProductosController.java
@@ -1,5 +1,0 @@
-package com.uppoteam.ecomercemariaharo.controller;
-
-public class ProductosController {
-
-}
