--- ecomercemariaharo/src/main/java/com/uppoteam/ecomercemariaharo/controller/UsuariosController.java
@@ -1,5 +1,0 @@
-package com.uppoteam.ecomercemariaharo.controller;
-
-public class UsuariosController {
-
-}
