--- ecomercemariaharo/src/main/java/com/uppoteam/ecomercemariaharo/service/PedidosService.java
@@ -1,10 +1,0 @@
-package com.uppoteam.ecomercemariaharo.service;
-
-public class PedidosService {
-	
-	
-	
-	
-	
-	
-}//PedidosService
