--- ecomercemariaharo/src/main/java/com/uppoteam/ecomercemariaharo/service/ProductosService.java
@@ -1,8 +1,0 @@
-package com.uppoteam.ecomercemariaharo.service;
-
-public class ProductosService {
-	
-	
-	
-
-}//ProductosService
