--- ecomercemariaharo/src/main/java/com/uppoteam/ecomercemariaharo/service/UsuariosService.java
@@ -1,5 +1,0 @@
-package com.uppoteam.ecomercemariaharo.service;
-
-public class UsuariosService {
-
-}
