+++ ecommercemariaharo/src/main/java/com/uppoteam/ecommercemariaharo/controller/PedidosController.java
@@ -1,0 +1,5 @@
+package com.uppoteam.ecommercemariaharo.controller;
+
+public class PedidosController {
+
+}
