+++ ecommercemariaharo/src/main/java/com/uppoteam/ecommercemariaharo/service/UsuariosService.java
@@ -1,0 +1,5 @@
+package com.uppoteam.ecommercemariaharo.service;
+
+public class UsuariosService {
+
+}
